requirements.txt
pyyaml>=6.0.1
pydantic>=2.8.0
click>=8.1.7
rich>=13.7.1
SQLAlchemy>=2.0
fastapi>=0.112.0
uvicorn>=0.30.0


